backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "aifactory-backend"
version = "1.0.0"
description = "AI Factory Prompt Assistant Backend - FastAPI + LangGraph"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn>=0.41.0",
    "langgraph>=1.1.0",
    "langchain-openai>=1.1.10",
    "langgraph-checkpoint-postgres>=3.0.4",
    "sse-starlette>=3.2.0",
    "pydantic-settings>=2.13.0",
    "psycopg[binary]>=3.3.3",
    "Pillow>=12.1.0",
    "langsmith>=0.7.16",
    "httpx>=0.27.0",
    "tavily-python>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
